pyproject.toml
[project]
name = "dataworkflow"
version = "0.1.0"
description = "Distributed workflow execution engine with version control"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "sqlalchemy>=2.0.36",
    "boto3>=1.34.17",
    "python-dotenv>=1.0.0",
    "werkzeug>=3.0.1",
    "markdown>=3.5.1",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
dataworkflow = "cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
]
postgres = [
    "psycopg2-binary>=2.9.9",
]

[build-system]
requires = ["setuptools>=61.0", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src", "sdk"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
